+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-resourcemanager/src/main/java/org/apache/hadoop/yarn/server/resourcemanager/scheduler/ClusterNodeTracker.java
@@ -57,7 +57,7 @@ public class ClusterNodeTracker<N extends SchedulerNode> {
 
   // Max allocation
   private long maxNodeMemory = -1;
-  private long maxNodeVCores = -1;
+  private int maxNodeVCores = -1;
   private Resource configuredMaxAllocation;
   private boolean forceConfiguredMaxAllocation = true;
   private long configuredMaxAllocationWaitTime;
